Struct sigstore::cosign::Client

pub struct Client { /* fields omitted */ }

Cosign Client

Given a container image/oci artifact, calculate the location of its cosign signature inside of a registry:

use crate::sigstore::cosign::CosignCapabilities;
use std::io::prelude::*;
use std::fs::File;

#[tokio::main]
pub async fn main() {
  let auth = &sigstore::registry::Auth::Anonymous;

  // Read the rekor key from the location generated by `cosign init`
  let mut rekor_pub_key_file = File::open("~/.sigstore/root/targets/rekor.pub")
    .expect("Cannot read rekor public key");
  let mut rekor_pub_key = String::new();
  rekor_pub_key_file.read_to_string(&mut rekor_pub_key)
    .expect("Not a ECDSA PEM encoded key");

  let mut client = sigstore::cosign::ClientBuilder::default()
    .with_rekor_pub_key(&rekor_pub_key)
    .build()
    .expect("Unexpected failure while building Client");
  let image = "registry-testing.svc.lan/kubewarden/disallow-service-nodeport:v0.1.0";
  let (cosign_signature_image, source_image_digest) = client.triangulate(
    image,
    auth
  ).await.expect("Unexpected failure while using triangulate");
}

Verify the signature of a container image/oci artifact:

use crate::sigstore::cosign::CosignCapabilities;
use std::collections::HashMap;
use std::io::prelude::*;
use std::fs::File;

#[tokio::main]
pub async fn main() {
  let auth = &sigstore::registry::Auth::Anonymous;

  // Read the rekor key from the location generated by `cosign init`
  let mut rekor_pub_key_file = File::open("~/.sigstore/root/targets/rekor.pub")
    .expect("Cannot read rekor public key");
  let mut rekor_pub_key = String::new();
  rekor_pub_key_file.read_to_string(&mut rekor_pub_key)
    .expect("Not a ECDSA PEM encoded key");

  let mut client = sigstore::cosign::ClientBuilder::default()
    .with_rekor_pub_key(&rekor_pub_key)
    .build()
    .expect("Unexpected failure while building Client");

  // Obtained via `triangulate`
  let cosign_image = "registry-testing.svc.lan/kubewarden/disallow-service-nodeport:sha256-5f481572d088dc4023afb35fced9530ced3d9b03bf7299c6f492163cb9f0452e.sig";
  // Obtained via `triangulate`
  let source_image_digest = "sha256-5f481572d088dc4023afb35fced9530ced3d9b03bf7299c6f492163cb9f0452e";

  let mut annotations: HashMap<String, String> = HashMap::new();
  annotations.insert("env".to_string(), "prod".to_string());

  let verification_key = "contents of a `cosign.pub` key read from the disk";
  let signatures_matching_requirements = client.verify(
    auth,
    cosign_image,
    source_image_digest,
    verification_key,
    Some(annotations)
  ).await.expect("unexpected verification error");

  if signatures_matching_requirements.is_empty() {
    panic!("no signature is matching the requirments");
  } else {
    println!("signatures matching the requirements: {:?}",
        signatures_matching_requirements);
  }
}