Module sigstore::cosign::verification_constraint

Structs that can be used to verify crate::cosign::SignatureLayer with special business logic.

This module provides already the most common kind of verification constraints:

PublicKeyVerifier: ensure a signature has been produced by a specific cosign key

CertSubjectEmailVerifier: ensure a signature has been produced in keyless mode, plus the email address associated with the signer matches a specific one

CertSubjectUrlVerifier: ensure a signature has been produced in keyless mode, plus the certificate SAN has a specific URI inside of it. This can be used to verify signatures produced by GitHub Actions.

Developers can define ad-hoc validation logic by creating a Struct that implements the VerificationConstraintVec trait.

Structs

AnnotationVerifier

Verification Constraint for the annotations added by cosign sign

CertSubjectEmailVerifier

Verification Constraint for signatures produced in keyless mode.

CertSubjectUrlVerifier

Verification Constraint for signatures produced in keyless mode.

PublicKeyVerifier

Verification Constraint for signatures produced with public/private keys

Traits

VerificationConstraint

A trait that can be used to define verification constraints objects that use a custom verification logic.

Type Definitions

VerificationConstraintVec

A list of objects implementing the VerificationConstraint trait