Trait sigstore::cosign::verification_constraint::VerificationConstraint

pub trait VerificationConstraint: Debug {
    fn verify(&self, signature_layer: &SignatureLayer) -> Result<bool>;
}

A trait that can be used to define verification constraints objects that use a custom verification logic.

Required methods

fn verify(&self, signature_layer: &SignatureLayer) -> Result<bool>

Given the signature_layer object, return true if the verification check is satisfied.

Developer can use the errors::SigstoreError::VerificationConstraintError error when something goes wrong inside of the verification logic.

use sigstore::{
  cosign::verification_constraint::VerificationConstraint,
  cosign::signature_layers::SignatureLayer,
  errors::{SigstoreError, Result},
};

#[derive(Debug)]
struct MyVerifier{}

impl VerificationConstraint for MyVerifier {
  fn verify(&self, _sl: &SignatureLayer) -> Result<bool> {
    Err(SigstoreError::VerificationConstraintError(
        "something went wrong!".to_string()))
  }
}