Module sigstore::tuf

Helper Structs to interact with the Sigstore TUF repository.

The main interaction point is SigstoreRepository, which fetches Rekor’s public key and Fulcio’s certificate.

These can later be given to cosign::ClientBuilder to enable Fulcio and Rekor integrations.

Example

The SigstoreRepository instance can be created via the SigstoreRepository::fetch method.

use sigstore::tuf::SigstoreRepository;
use sigstore::cosign;

let repo = SigstoreRepository::fetch(None)
    .expect("Error while building SigstoreRepository");
let client = cosign::ClientBuilder::default()
    .with_rekor_pub_key(repo.rekor_pub_key())
    .with_fulcio_cert(repo.fulcio_cert())
    .build()
    .expect("Error while building cosign client");

The SigstoreRepository::fetch method can attempt to leverage local copies of the Rekor and Fulcio files. Please refer to the method docs for more details.

Warning: the SigstoreRepository::fetch method currently needs special handling when invoked inside of an async context. Please refer to the method docs for more details.

Structs

SigstoreRepository

Securely fetches Rekor public key and Fulcio certificate from Sigstore’s TUF repository